refactor(web): extract runs router payload helpers to _runs_payloads
Moves the five pure dict/event shaping helpers out of the 1809-LOC
amx/web/routers/runs.py into a focused
amx/web/routers/_runs_payloads.py:

- _column_details_for_table (SSE per-table row hydration from history)
- _emit_tokens_snapshot (running token+cost snapshot for SSE stream)
- _review_result_to_event (compact ReviewResult dict for activity.complete)
- _resolve_apply_results (apply-request body or on-disk pending queue)
- _review_result_from_dict (dict -> ReviewResult coercion for POST body)

runs.py re-exports each name. _resolve_apply_results reads
load_pending off the runs module (lazy import) so existing tests that
monkeypatch runs.load_pending keep affecting the call site.

runs.py: 1809 -> 1658 LOC (-151, -8%).
